+++ scout/server/blueprints/cases/views.py
@@ -293,6 +293,16 @@ def pdf_case_report(institute_id, case_name):
     )
 
 
+@cases_bp.route("/<institute_id>/<case_name>/chanjo2_coverage_report", methods=["GET"])
+async def chanjo2_coverage_report(institute_id, case_name) -> str:
+    """Redirect to a Chanjo2 coverage report."""
+
+    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    return await controllers.chanjo2_coverage_report_content(
+        institute_obj=institute_obj, case_obj=case_obj, panel_name=request.args.get("panel_name")
+    )
+
+
 @cases_bp.route("/<institute_id>/<case_name>/mt_report", methods=["GET"])
 def mt_report(institute_id, case_name):
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)